Style Changes #40
- remove the history button from the filter options in kompakkt and add it to the navigation bar like in semantic kompakkt
 --> VE: Removed login and registration button from actionbar, removed actionbar from home-page
- add the New Object and for now for kompakkt also the New Collections button to the navigation bar
- change the fontsize from the button to 14
- if possible change also the search options to font size 14
- in semantic kompakkt it also says "Search for objects and compilations", but we don't have the "compilations" here, only in kompakkt, so better write here "Search in objects" and in kompakkt better "Search in objects and collections" (or depending from the toggle function of the Objects and Collections)
- align the search options with the logo at the navigation bar
- gap between buttons, history button and "Profile"/"Logout": 8px
