+++ src/app/components/actionbar/actionbar.component.scss
@@ -3,7 +3,9 @@
 #actionbar {
   width: 100%;
   position: sticky;
-  font-size: 12px;
+  font-size: 14px;
+
+  padding-left: 30px;
 
   background-color: #fff;
   border-bottom: 2px solid #eee;
@@ -13,7 +15,7 @@
 
   display: flex;
   flex-direction: row;
-  justify-content: center;
+  justify-content: start;
 
   @include queries.less-than-large() {
     display: none;
@@ -28,20 +30,9 @@
   margin-right: 15px;
 }
 
-.history-button {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-
-  p {
-    margin: 0;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-}
-
 .container {
   padding-top: 10px;
+  
 }
 
 .quick-add-button {
@@ -53,4 +44,9 @@
   mat-icon {
     margin-right: 0.3rem;
   }
+
+  .mdc-button__label {
+    text-align: start;
+  }
 }
+
