+++ src/app/pages/explore/explore.component.scss
@@ -78,6 +78,9 @@ mat-option {
   font-weight: 400;
   mat-icon {
     margin-right: 0.3rem;
+    height: auto;
+    width: auto;
+    font-size: 24px;
   }
   * {
     padding-right: 0.4rem;
