+++ src/app/pages/home/home.component.html
@@ -55,7 +55,7 @@
     </div>
   </div>
 
-  <app-actionbar></app-actionbar>
+  <!-- <app-actionbar></app-actionbar> -->
 
   <section class="content" id="kompakkt-previews">
     <div class="entity-grid">
