+++ src/assets/i18n/en.json
@@ -177,7 +177,7 @@
   "Upload capabilities requested": "Upload capabilities requested",
   "New Object": "New Object",
   "New Collection": "New Collection",
-  "Search in Objects & Compilations.": "Search in Objects & Compilations.",
+  "Search in Objects & Collections.": "Search in Objects & Collections",
   "Objects": "Objects",
   "Collections": "Collections",
   "Media Types": "Media Types",
