+++ src/styles.scss
@@ -415,12 +415,6 @@ app-actionbar {
       padding: 0 0.4rem;
     }
   }
-
-  button {
-    font-size: 12px;
-    margin: 0 0.4rem;
-    border-radius: 25px;
-  }
 }
 
 /* Scrollbar style */
@@ -567,6 +561,7 @@ app-annotate {
 }
 
 #actionbar {
+
   .mat-mdc-form-field {
     line-height: 1.3 !important;
   }
@@ -604,6 +599,15 @@ app-annotate {
   .mdc-floating-label {
     text-overflow: ellipsis !important;
     width: 180px !important;
+    font-size: 14px;
+  }
+
+  .mat-mdc-select-value-text {
+    font-size: 14px;
+  }
+
+  .filterTypes {
+    font-size: 14px;
   }
 }
 
@@ -621,6 +625,13 @@ app-annotate {
    display: none;
  }
 
+}
+
+.mat-mdc-menu-panel {
+  max-width: 350px !important;
+}
+.mat-mdc-button {
+  letter-spacing: normal !important;
 }
 
 /* Dialogs */
